+++ tests/test_tables.py
@@ -4,7 +4,7 @@ from .context import test_dir
 
 # Manual test (requires inspection of result) for converting html with nested tables
 
-filename = os.path.join(f'{test_dir}/assets/htmls', 'tables2.html')
+filename = os.path.join(f'{test_dir}/assets/htmls', 'tables3.html')
 d = HtmlToDocx()
 
-d.parse_html_file(filename, f'{test_dir}/tables2')
+d.parse_html_file(filename, f'{test_dir}/tables3')